requirements.txt
flask
requests
pyOpenSSL
zeep
cryptography
sqlalchemy
flask-login
pyyaml
